Add new file writeProperty.go for writing a BACnet property value
This commit adds a new file writeProperty.go which contains an example main function for writing a BACnet property value. The function sets up the necessary variables and creates the NPDU (Network Protocol Data Unit) for sending the request to the BACnet device. It then encodes the NPDU and creates the APDU (Application Protocol Data Unit) with the necessary information for a WriteProperty service request.

The purpose of this commit is to provide the functionality for writing a BACnet property value to a BACnet device.

Signed-off-by: SammyOina <sammyoina@gmail.com>